+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Innmind\BlackBox\PHPUnit\Load::parseTagWith()`
 - `Innmind\BlackBox\Application::stopOnFailure()`
 - `Innmind\BlackBox\Application::map()`
+- `Innmind\BlackBox\Application::when()`
 
 ## 5.4.0 - 2023-09-02